refactor: modularize VS Code extension and update dependencies
- Split 1220-line extension.ts into 9 focused modules
- Update TypeScript to v5 and other dev dependencies
- Add JSDoc documentation to all modules
- Fix unused imports in extension.ts and diagnostics.ts
