+++ vscode-extension/mtlog-analyzer/src/extension.ts
@@ -1,339 +1,48 @@
 import * as vscode from 'vscode';
-import { spawn, ChildProcess, execSync } from 'child_process';
-import * as path from 'path';
-import * as os from 'os';
-import * as crypto from 'crypto';
-import * as fs from 'fs';
+import { VSCodeMtlogLogger, logWithTimestamp } from './logger';
+import { registerCommands } from './commands';
+import { checkAnalyzerAvailable } from './analyzer';
+import { 
+    initializeDiagnostics, 
+    queueAnalysis, 
+    cleanupProcesses,
+    getRunningAnalyses,
+    getAnalysisQueueLength
+} from './diagnostics';
+import { initializeStatusBar, updateStatusBar } from './statusBar';
+import { MtlogCodeActionProvider } from './codeActions';
 
-// Constants
-const LOG_TRUNCATION_LENGTH = 200;  // Maximum length for log output in debug messages
-const STATUS_BAR_WIDTH = 23;         // Fixed width for status bar to prevent jumping
-const ENV_MTLOG_SUPPRESS = 'MTLOG_SUPPRESS';  // Environment variable for diagnostic suppression
-const ENV_MTLOG_DISABLE_ALL = 'MTLOG_DISABLE_ALL';  // Environment variable to disable all diagnostics
-
-// Logging utility for consistent timestamp formatting
-function logWithTimestamp(channel: vscode.OutputChannel, message: string): void {
-    channel.appendLine(`[${new Date().toLocaleTimeString()}] ${message}`);
-}
-
-// Enhanced logger similar to GoLand's MtlogLogger
-class VSCodeMtlogLogger {
-    private static instance: VSCodeMtlogLogger;
-    private outputChannel: vscode.OutputChannel;
-    
-    private constructor(outputChannel: vscode.OutputChannel) {
-        this.outputChannel = outputChannel;
-    }
-    
-    static initialize(outputChannel: vscode.OutputChannel): void {
-        VSCodeMtlogLogger.instance = new VSCodeMtlogLogger(outputChannel);
-    }
-    
-    static debug(message: string, showInUI: boolean = false): void {
-        if (VSCodeMtlogLogger.instance) {
-            VSCodeMtlogLogger.instance.outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] DEBUG: ${message}`);
-            if (showInUI) {
-                vscode.window.showInformationMessage(`[DEBUG] ${message}`);
-            }
-        }
-    }
-    
-    static info(message: string, showInUI: boolean = false): void {
-        if (VSCodeMtlogLogger.instance) {
-            VSCodeMtlogLogger.instance.outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] INFO: ${message}`);
-            if (showInUI) {
-                vscode.window.showInformationMessage(message);
-            }
-        }
-    }
-    
-    static warn(message: string, showInUI: boolean = true): void {
-        if (VSCodeMtlogLogger.instance) {
-            VSCodeMtlogLogger.instance.outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] WARN: ${message}`);
-            if (showInUI) {
-                vscode.window.showWarningMessage(message);
-            }
-        }
-    }
-    
-    static error(message: string, error?: Error, showInUI: boolean = true): void {
-        if (VSCodeMtlogLogger.instance) {
-            const errorDetails = error ? ` | Error: ${error.message}\nStack: ${error.stack}` : '';
-            VSCodeMtlogLogger.instance.outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] ERROR: ${message}${errorDetails}`);
-            if (showInUI) {
-                vscode.window.showErrorMessage(message);
-            }
-        }
-    }
-    
-    static show(): void {
-        if (VSCodeMtlogLogger.instance) {
-            VSCodeMtlogLogger.instance.outputChannel.show();
-        }
-    }
-}
-
-// Extended diagnostic interface to store analyzer suggested fixes
-interface MtlogDiagnostic extends vscode.Diagnostic {
-    // Store suggested fixes from analyzer
-    suggestedFixes?: Array<{
-        message: string;
-        textEdits: Array<{
-            pos: string;
-            end: string;
-            newText: string;
-        }>;
-    }>;
-}
-
-let diagnosticCollection: vscode.DiagnosticCollection;
-let outputChannel: vscode.OutputChannel;
-let statusBarItem: vscode.StatusBarItem;
-let activeProcesses = new Map<string, ChildProcess>();
-let analysisVersions = new Map<string, number>();
-let analysisQueue: string[] = [];
-let runningAnalyses = 0;
-const maxConcurrentAnalyses = Math.max(1, os.cpus().length - 1);
-let totalIssueCount = 0;
-
-// Cache removed - was causing stale diagnostics to persist
-
+/**
+ * Extension activation point.
+ * Sets up all modules, registers handlers, and starts initial analysis.
+ */
 export function activate(context: vscode.ExtensionContext) {
     // Create output channel for error logging and diagnostics
-    outputChannel = vscode.window.createOutputChannel('mtlog-analyzer');
+    const outputChannel = vscode.window.createOutputChannel('mtlog-analyzer');
     context.subscriptions.push(outputChannel);
     
     // Initialize enhanced logger
     VSCodeMtlogLogger.initialize(outputChannel);
     VSCodeMtlogLogger.info('mtlog-analyzer extension activated');
     
-    diagnosticCollection = vscode.languages.createDiagnosticCollection('mtlog');
+    // Initialize diagnostic collection
+    const diagnosticCollection = vscode.languages.createDiagnosticCollection('mtlog');
     context.subscriptions.push(diagnosticCollection);
     
-    // Create status bar item - clicking toggles Problems panel
-    statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 100);
-    statusBarItem.command = 'workbench.actions.view.toggleProblems';
-    context.subscriptions.push(statusBarItem);
-    updateStatusBar();
+    // Initialize diagnostics module
+    initializeDiagnostics(diagnosticCollection, outputChannel);
     
-    // Register command to show output
-    context.subscriptions.push(
-        vscode.commands.registerCommand('mtlog.showOutput', () => {
-            showDiagnosticsSummary();
-            outputChannel.show();
-        })
-    );
+    // Initialize status bar
+    initializeStatusBar(context);
     
-    // Register manual analysis command
-    context.subscriptions.push(
-        vscode.commands.registerCommand('mtlog.analyzeNow', async () => {
-            const document = vscode.window.activeTextEditor?.document;
-            if (document && document.languageId === 'go') {
-                logWithTimestamp(outputChannel, `Manual analysis triggered for ${document.fileName}`);
-                diagnosticCollection.delete(document.uri);
-                await analyzeDocument(document);
-            }
-        })
-    );
-    
-    // Register save and analyze command (for quick fixes)
-    context.subscriptions.push(
-        vscode.commands.registerCommand('mtlog.saveAndAnalyze', async () => {
-            const document = vscode.window.activeTextEditor?.document;
-            if (document && document.languageId === 'go') {
-                logWithTimestamp(outputChannel, `Save and analyze triggered for ${document.fileName}`);
-                await document.save();
-                // Small delay to ensure file is saved
-                await new Promise(resolve => setTimeout(resolve, 100));
-                diagnosticCollection.delete(document.uri);
-                await analyzeDocument(document);
-            }
-        })
-    );
-    
-    // Register toggle diagnostics command
-    context.subscriptions.push(
-        vscode.commands.registerCommand('mtlog.toggleDiagnostics', async () => {
-            const config = vscode.workspace.getConfiguration('mtlog');
-            const currentState = config.get<boolean>('diagnosticsEnabled', true);
-            await config.update('diagnosticsEnabled', !currentState, vscode.ConfigurationTarget.Workspace);
-            
-            const newState = !currentState ? 'enabled' : 'disabled';
-            vscode.window.showInformationMessage(`mtlog-analyzer ${newState}`);
-            logWithTimestamp(outputChannel, `Diagnostics ${newState}`);
-            
-            // Re-analyze all open files
-            diagnosticCollection.clear();
-            vscode.workspace.textDocuments.forEach(document => {
-                if (document.languageId === 'go') {
-                    queueAnalysis(document);
-                }
-            });
-        })
-    );
-    
-    // Register suppress diagnostic command
-    context.subscriptions.push(
-        vscode.commands.registerCommand('mtlog.suppressDiagnostic', async (diagnosticId?: string) => {
-            logWithTimestamp(outputChannel, `Suppress diagnostic command called with: ${diagnosticId || 'no ID'}`);
-            
-            // If we got a URI instead of a diagnostic ID, ignore it
-            if (diagnosticId && (diagnosticId.startsWith('file://') || diagnosticId.includes('://'))) {
-                logWithTimestamp(outputChannel, `Ignoring URI passed as diagnostic ID: ${diagnosticId}`);
-                diagnosticId = undefined;
-            }
-            
-            if (!diagnosticId) {
-                // Try to get diagnostic ID from current cursor position
-                const editor = vscode.window.activeTextEditor;
-                if (editor) {
-                    const diagnostics = diagnosticCollection.get(editor.document.uri);
-                    logWithTimestamp(outputChannel, `Found ${diagnostics?.length || 0} diagnostics in current file`);
-                    
-                    if (diagnostics) {
-                        const position = editor.selection.active;
-                        const diagnostic = diagnostics.find(d => d.range.contains(position));
-                        
-                        if (diagnostic) {
-                            logWithTimestamp(outputChannel, `Found diagnostic at cursor: ${diagnostic.message}`);
-                            logWithTimestamp(outputChannel, `Diagnostic code: ${(diagnostic as any).code}`);
-                            logWithTimestamp(outputChannel, `Diagnostic full object: ${JSON.stringify(diagnostic)}`);
-                            
-                            // First try to use the code property
-                            if ((diagnostic as any).code) {
-                                diagnosticId = String((diagnostic as any).code);
-                                outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Using code property: ${diagnosticId}`);
-                            } else if (diagnostic.message) {
-                                // Fallback to extracting from message
-                                const match = diagnostic.message.match(/\[?(MTLOG\d{3})\]?/);
-                                if (match) {
-                                    diagnosticId = match[1];
-                                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Extracted from message: ${diagnosticId}`);
-                                }
-                            }
-                        } else {
-                            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] No diagnostic found at cursor position`);
-                        }
-                    }
-                }
-                
-                if (!diagnosticId) {
-                    // Show output to help debug
-                    outputChannel.show();
-                    diagnosticId = await vscode.window.showInputBox({
-                        prompt: 'Enter diagnostic ID to suppress (e.g., MTLOG001)',
-                        placeHolder: 'MTLOG001'
-                    });
-                }
-            }
-            
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Will suppress diagnostic ID: ${diagnosticId || 'none'}`);
-            
-            if (diagnosticId) {
-                const config = vscode.workspace.getConfiguration('mtlog');
-                
-                // Get current suppressed list (or empty array)
-                const currentValue = config.inspect('suppressedDiagnostics');
-                outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Current config value: ${JSON.stringify(currentValue)}`);
-                
-                // Try to get existing valid IDs
-                const existing = config.get('suppressedDiagnostics');
-                let suppressed: string[] = [];
-                
-                if (existing === undefined || existing === null) {
-                    // No config exists yet, start with empty array
-                    suppressed = [];
-                } else if (Array.isArray(existing)) {
-                    // Valid array, filter for valid IDs
-                    for (const item of existing) {
-                        if (typeof item === 'string' && item.startsWith('MTLOG')) {
-                            suppressed.push(item);
-                        }
-                    }
-                } else {
-                    // Config is corrupted, abort to avoid data loss
-                    vscode.window.showErrorMessage('Failed to parse suppressed diagnostics configuration. Suppression aborted to avoid losing existing suppressed diagnostics.');
-                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] ERROR: suppressedDiagnostics config is not a valid array: ${typeof existing}. Aborting suppression.`);
-                    outputChannel.show();
-                    return;
-                }
-                
-                if (!suppressed.includes(diagnosticId)) {
-                    suppressed.push(diagnosticId);
-                    
-                    // Clear first, then set
-                    await config.update('suppressedDiagnostics', undefined, vscode.ConfigurationTarget.Workspace);
-                    await config.update('suppressedDiagnostics', suppressed, vscode.ConfigurationTarget.Workspace);
-                    
-                    vscode.window.showInformationMessage(`Suppressed diagnostic ${diagnosticId}`);
-                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Saved suppressed list: [${suppressed.join(', ')}]`);
-                    
-                    // Re-analyze to apply suppression
-                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Clearing all diagnostics and re-analyzing...`);
-                    diagnosticCollection.clear();
-                    
-                    const goDocuments = vscode.workspace.textDocuments.filter(d => d.languageId === 'go');
-                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Found ${goDocuments.length} Go files to re-analyze`);
-                    
-                    goDocuments.forEach(document => {
-                        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Queueing re-analysis for ${document.fileName}`);
-                        queueAnalysis(document);
-                    });
-                } else {
-                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] ${diagnosticId} already in suppressed list`);
-                }
-            }
-        })
-    );
-    
-    // Register suppression manager command
-    context.subscriptions.push(
-        vscode.commands.registerCommand('mtlog.showSuppressionManager', async () => {
-            const config = vscode.workspace.getConfiguration('mtlog');
-            const suppressed = config.get<string[]>('suppressedDiagnostics', []);
-            
-            if (suppressed.length === 0) {
-                vscode.window.showInformationMessage('No diagnostics are currently suppressed');
-                return;
-            }
-            
-            const items = suppressed.map(id => ({
-                label: id,
-                description: getDiagnosticDescription(id),
-                picked: true
-            }));
-            
-            const selected = await vscode.window.showQuickPick(items, {
-                canPickMany: true,
-                placeHolder: 'Select diagnostics to keep suppressed (uncheck to unsuppress)'
-            });
-            
-            if (selected !== undefined) {
-                const newSuppressed = selected.map(item => item.label);
-                await config.update('suppressedDiagnostics', newSuppressed, vscode.ConfigurationTarget.Workspace);
-                
-                const removed = suppressed.filter(id => !newSuppressed.includes(id));
-                if (removed.length > 0) {
-                    vscode.window.showInformationMessage(`Unsuppressed: ${removed.join(', ')}`);
-                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Unsuppressed: ${removed.join(', ')}`);
-                    
-                    // Re-analyze to apply changes
-                    diagnosticCollection.clear();
-                    vscode.workspace.textDocuments.forEach(document => {
-                        if (document.languageId === 'go') {
-                            queueAnalysis(document);
-                        }
-                    });
-                }
-            }
-        })
-    );
+    // Register all commands
+    registerCommands(context);
 
     // Register save handler for immediate analysis
     context.subscriptions.push(
         vscode.workspace.onDidSaveTextDocument(document => {
             if (document.languageId === 'go') {
-                outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Save triggered analysis for ${document.fileName}`);
+                logWithTimestamp(outputChannel, `Save triggered analysis for ${document.fileName}`);
                 queueAnalysis(document);
             }
         })
@@ -377,13 +86,8 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(
         vscode.workspace.onDidChangeConfiguration(event => {
             if (event.affectsConfiguration('mtlog')) {
-                // Cache removed - no longer needed
-                
                 // Cancel all active processes
-                for (const [file, proc] of activeProcesses) {
-                    proc.kill();
-                }
-                activeProcesses.clear();
+                cleanupProcesses();
                 
                 // Re-analyze all open Go files
                 vscode.workspace.textDocuments.forEach(document => {
@@ -405,816 +109,21 @@ export function activate(context: vscode.ExtensionContext) {
             }
         )
     );
+    
+    // Update status bar periodically
+    const statusBarUpdateInterval = setInterval(() => {
+        updateStatusBar(diagnosticCollection, getRunningAnalyses(), getAnalysisQueueLength());
+    }, 1000);
+    
+    context.subscriptions.push({
+        dispose: () => clearInterval(statusBarUpdateInterval)
+    });
 }
 
+/**
+ * Extension deactivation point.
+ * Cleans up running analyzer processes.
+ */
 export function deactivate() {
-    // Kill all active processes
-    for (const [file, proc] of activeProcesses) {
-        proc.kill();
-    }
-    activeProcesses.clear();
-    analysisQueue = [];
-    
-    if (statusBarItem) {
-        statusBarItem.hide();
-    }
-}
-
-function queueAnalysis(document: vscode.TextDocument) {
-    const filePath = document.fileName;
-    
-    // Remove from queue if already present
-    const index = analysisQueue.indexOf(filePath);
-    if (index !== -1) {
-        analysisQueue.splice(index, 1);
-    }
-    
-    // Add to end of queue
-    analysisQueue.push(filePath);
-    
-    // Process queue if we have capacity
-    processQueue();
-}
-
-async function processQueue() {
-    while (analysisQueue.length > 0 && runningAnalyses < maxConcurrentAnalyses) {
-        const filePath = analysisQueue.shift();
-        if (!filePath) continue;
-        
-        const document = vscode.workspace.textDocuments.find(d => d.fileName === filePath);
-        if (document) {
-            runningAnalyses++;
-            updateStatusBar();
-            analyzeDocument(document).finally(() => {
-                runningAnalyses--;
-                updateStatusBar();
-                processQueue(); // Process next in queue
-            });
-        }
-    }
-}
-
-async function analyzeDocument(document: vscode.TextDocument) {
-    const filePath = document.fileName;
-    
-    // Increment analysis version for this file
-    const currentVersion = (analysisVersions.get(filePath) || 0) + 1;
-    analysisVersions.set(filePath, currentVersion);
-    
-    // Clear existing diagnostics immediately to prevent phantom diagnostics
-    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] CLEARING diagnostics for ${document.uri.fsPath} (version ${currentVersion})`);
-    diagnosticCollection.delete(document.uri);
-    
-    // Cancel any existing analysis for this file
-    const existingProcess = activeProcesses.get(filePath);
-    if (existingProcess) {
-        existingProcess.kill();
-        activeProcesses.delete(filePath);
-    }
-    
-    // We need to always re-analyze to detect when problems are fixed
-    // Don't use cache - it causes stale diagnostics to persist after fixes
-    const fileContent = document.getText();
-    
-    const diagnostics: vscode.Diagnostic[] = [];
-    const fileUri = document.uri;
-    let fileHash = '';
-    
-    // Find the Go module root
-    let workingDir = path.dirname(document.fileName);
-    let packagePath = './...';
-    try {
-        const goModPath = execSync('go env GOMOD', { 
-            cwd: workingDir, 
-            encoding: 'utf8' 
-        }).trim();
-        
-        if (goModPath && goModPath !== 'NUL' && goModPath !== '/dev/null') {
-            workingDir = path.dirname(goModPath);
-            // Get relative path from module root to current file's directory
-            const relPath = path.relative(workingDir, path.dirname(document.fileName));
-            packagePath = relPath ? `./${relPath.replace(/\\/g, '/')}/...` : './...';
-        }
-    } catch (e) {
-        // Fall back to single file if not in a module
-    }
-    
-    // Not using cache anymore - causes stale diagnostics
-    
-    // Get the analyzer path using smart detection
-    const analyzerPath = await findAnalyzer();
-    if (!analyzerPath) {
-        VSCodeMtlogLogger.error('Analyzer not found - cannot run analysis', undefined, false);
-        await checkAnalyzerAvailable();
-        return;
-    }
-    
-    // Get flags from config
-    const config = vscode.workspace.getConfiguration('mtlog');
-    const analyzerFlags = config.get<string[]>('analyzerFlags', []);
-    
-    // Add kill switch flags based on configuration
-    const diagnosticsEnabled = config.get<boolean>('diagnosticsEnabled', true);
-    let suppressedDiagnostics = config.get('suppressedDiagnostics', []);
-    
-    // Debug: Log raw config value
-    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Raw suppressedDiagnostics from config: ${JSON.stringify(suppressedDiagnostics)}`);
-    
-    // Fix: Ensure suppressed diagnostics are strings, not objects or URIs
-    // Handle both array and single value cases  
-    let suppressedArray: string[] = [];
-    if (Array.isArray(suppressedDiagnostics)) {
-        suppressedArray = suppressedDiagnostics
-            .map((d: any) => {
-                if (typeof d === 'string') {
-                    // Filter out URIs that got saved by mistake
-                    if (!d.includes('://') && d.startsWith('MTLOG')) {
-                        return d;
-                    }
-                }
-                // Ignore objects (like URI objects)
-                return null;
-            })
-            .filter((d): d is string => d !== null);
-    }
-    
-    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Diagnostics enabled: ${diagnosticsEnabled}`);
-    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Processed suppressed diagnostics: [${suppressedArray.join(', ')}]`);
-    
-    // Use analyzer stdin mode to get suggested fixes
-    const args = ['-stdin'];
-    
-    // Add analyzer flags
-    args.push(...analyzerFlags);
-    
-    // Add suppression flags
-    if (!diagnosticsEnabled) {
-        args.push('-disable-all');
-    } else if (suppressedArray.length > 0) {
-        args.push(`-suppress=${suppressedArray.join(',')}`);
-    }
-    
-    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Running analyzer: ${analyzerPath} ${args.join(' ')}`);
-    if (suppressedArray.length > 0) {
-        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Suppressing: ${suppressedArray.join(',')}`);
-    }
-    
-    const proc = spawn(analyzerPath, args, {
-        cwd: workingDir
-    });
-    
-    // Track this process
-    activeProcesses.set(filePath, proc);
-    
-    // Send file content to stdin
-    const fileDocument = vscode.workspace.textDocuments.find(doc => doc.uri.fsPath === filePath);
-    if (fileDocument) {
-        const stdinRequest = {
-            filename: filePath,
-            content: fileDocument.getText(),
-            go_module: workingDir
-        };
-        
-        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Sending ${stdinRequest.content.length} chars to analyzer stdin`);
-        proc.stdin?.write(JSON.stringify(stdinRequest));
-        proc.stdin?.end();
-    }
-    
-    // Handle stdout - analyzer outputs JSON array of diagnostics
-    let stdoutBuffer = '';
-    proc.stdout.on('data', data => {
-        stdoutBuffer += data.toString();
-    });
-    
-    proc.stderr.on('data', (data) => {
-        const text = data.toString();
-        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] stderr: ${text.substring(0, LOG_TRUNCATION_LENGTH)}`);
-    });
-    
-    proc.on('close', (code) => {
-        // Clean up process tracking
-        activeProcesses.delete(filePath);
-        
-        // Only apply diagnostics if this is the latest analysis version
-        const latestVersion = analysisVersions.get(filePath) || 0;
-        if (currentVersion < latestVersion) {
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Discarding stale analysis results (version ${currentVersion} < ${latestVersion})`);
-            return;
-        }
-        
-        // Parse JSON array from stdout
-        if (stdoutBuffer.trim()) {
-            try {
-                const stdinDiagnostics = JSON.parse(stdoutBuffer);
-                outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Parsed ${stdinDiagnostics.length} diagnostics from analyzer stdout`);
-                
-                // Convert analyzer diagnostics to VS Code diagnostics
-                for (const stdinDiag of stdinDiagnostics) {
-                    // Call pushDiag directly with the analyzer data
-                    const posn = `${stdinDiag.filename}:${stdinDiag.line}:${stdinDiag.column}`;
-                    pushDiagInternal(posn, stdinDiag.message, stdinDiag.severity, stdinDiag.suggestedFixes, stdinDiag.diagnostic_id, fileUri, diagnostics);
-                }
-            } catch (error) {
-                outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Failed to parse analyzer output: ${error}`);
-            }
-        }
-        
-        // Update diagnostics for this file
-        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] SETTING ${diagnostics.length} diagnostics for ${fileUri.fsPath} (version ${currentVersion})`);
-        for (const diag of diagnostics) {
-            outputChannel.appendLine(`  - Line ${diag.range.start.line + 1}: ${diag.message}`);
-        }
-        diagnosticCollection.set(fileUri, diagnostics);
-        
-        // Not caching anymore - it causes stale diagnostics to persist
-        
-        // Update total issue count
-        updateTotalIssueCount();
-        
-        // Log analysis completion
-        const relPath = vscode.workspace.asRelativePath(fileUri.fsPath);
-        if (diagnostics.length > 0) {
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Found ${diagnostics.length} issue${diagnostics.length !== 1 ? 's' : ''} in ${relPath}`);
-        } else {
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] No issues found in ${relPath} (exit code: ${code})`);
-        }
-    });
-    
-    proc.on('error', (err) => {
-        // Clean up process tracking
-        activeProcesses.delete(filePath);
-        
-        // Cache removed - no longer needed
-        
-        vscode.window.showErrorMessage(`mtlog-analyzer: ${err.message}`);
-    });
-}
-
-function pushDiagInternal(posn: string, message: string, category?: string, suggestedFixes?: any[], diagnosticId?: string, uri?: vscode.Uri, diagnostics?: vscode.Diagnostic[]): void {
-    if (!uri || !diagnostics) return;
-    
-    const posnParts = posn.split(':');
-    if (posnParts.length < 3) return;
-    
-    // Extract file, line, col
-    const file = posnParts.slice(0, -2).join(':');
-    const line = parseInt(posnParts[posnParts.length - 2]);
-    const col = parseInt(posnParts[posnParts.length - 1]);
-    
-    if (path.resolve(file) !== path.resolve(uri.fsPath)) return;
-    
-    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Raw diagnostic message: "${message}"`);
-    
-    // Parse severity from category or message prefix
-    let severity = vscode.DiagnosticSeverity.Error;
-    let cleanMessage = message;
-    
-    if (message.startsWith('warning:')) {
-        severity = vscode.DiagnosticSeverity.Warning;
-        cleanMessage = message.substring(8).trim();
-    } else if (message.startsWith('suggestion:')) {
-        severity = vscode.DiagnosticSeverity.Information;
-        cleanMessage = message.substring(11).trim();
-    } else if (message.startsWith('error:')) {
-        cleanMessage = message.substring(6).trim();
-    } else if (category) {
-        const cat = category.toLowerCase();
-        if (cat.includes('warn')) severity = vscode.DiagnosticSeverity.Warning;
-        else if (cat.includes('suggest') || cat.includes('info')) severity = vscode.DiagnosticSeverity.Information;
-    }
-    
-    // Validate line number exists in document
-    const document = vscode.workspace.textDocuments.find(d => d.uri.fsPath === uri.fsPath);
-    if (!document || line - 1 >= document.lineCount) {
-        // Log filtered diagnostic for debugging
-        if (outputChannel) {
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Filtered out-of-range diagnostic: line ${line} (doc has ${document?.lineCount || 0} lines)`);
-        }
-        return;
-    }
-    
-    // Use actual line length for end column
-    const lineText = document.lineAt(line - 1).text;
-    const diag = new vscode.Diagnostic(
-        new vscode.Range(line - 1, col - 1, line - 1, lineText.length),
-        cleanMessage,
-        severity
-    ) as MtlogDiagnostic;
-    
-    diag.source = 'mtlog';
-    
-    // Store suggested fixes from analyzer if available
-    if (suggestedFixes && suggestedFixes.length > 0) {
-        diag.suggestedFixes = suggestedFixes;
-        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Diagnostic has ${suggestedFixes.length} suggested fixes from analyzer`);
-    }
-    
-    // Use diagnostic ID from analyzer if provided
-    if (diagnosticId) {
-        (diag as any).code = diagnosticId;
-        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Set diagnostic code: ${diagnosticId} for message: ${cleanMessage.substring(0, 50)}...`);
-    }
-    
-    diagnostics.push(diag);
-}
-
-
-
-function updateStatusBar() {
-    if (!statusBarItem) return;
-    
-    const config = vscode.workspace.getConfiguration('mtlog');
-    const diagnosticsEnabled = config.get<boolean>('diagnosticsEnabled', true);
-    const suppressedDiagnostics = config.get<string[]>('suppressedDiagnostics', []);
-    
-    const errorCount = countIssuesBySeverity(vscode.DiagnosticSeverity.Error);
-    const warningCount = countIssuesBySeverity(vscode.DiagnosticSeverity.Warning);
-    const infoCount = countIssuesBySeverity(vscode.DiagnosticSeverity.Information);
-    
-    // Always use the same structure with non-breaking spaces (U+00A0) for consistent width
-    const e = errorCount || 0;
-    const w = warningCount || 0; 
-    const i = infoCount || 0;
-    const s = suppressedDiagnostics.length || 0;
-    
-    let tooltip = '';
-    let text = '';
-    
-    if (!diagnosticsEnabled) {
-        // When disabled, just show the disabled icon
-        text = '$(circle-slash)        '; // 17 chars + 6 spaces = STATUS_BAR_WIDTH total
-        tooltip = 'mtlog disabled';
-    } else if (runningAnalyses > 0 || analysisQueue.length > 0) {
-        // Show spinning icon during analysis
-        text = '$(sync~spin)           '; // 15 chars + 8 spaces = STATUS_BAR_WIDTH total  
-        tooltip = `mtlog: Analyzing ${runningAnalyses} file(s)`;
-    } else {
-        // Always show the 3 main icons with counts
-        const parts = [];
-        parts.push(`$(error) ${e}`);
-        parts.push(`$(warning) ${w}`);
-        parts.push(`$(lightbulb) ${i}`);
-        if (s > 0) parts.push(`$(eye-closed) ${s}`);
-        
-        const content = parts.join(' ');
-        text = content.padEnd(STATUS_BAR_WIDTH, ' '); // Always exactly STATUS_BAR_WIDTH characters
-        
-        // Build tooltip
-        const tooltipParts = [];
-        if (e > 0) tooltipParts.push(`${e} error${e !== 1 ? 's' : ''}`);
-        if (w > 0) tooltipParts.push(`${w} warning${w !== 1 ? 's' : ''}`);
-        if (i > 0) tooltipParts.push(`${i} suggestion${i !== 1 ? 's' : ''}`);
-        if (s > 0) tooltipParts.push(`${s} suppressed`);
-        
-        tooltip = tooltipParts.length > 0 
-            ? `mtlog: ${tooltipParts.join(', ')}`
-            : 'mtlog: No issues found';
-    }
-    
-    statusBarItem.text = text;
-    statusBarItem.tooltip = tooltip;
-    statusBarItem.show();
-}
-
-function updateTotalIssueCount() {
-    totalIssueCount = 0;
-    diagnosticCollection.forEach((uri, diagnostics) => {
-        totalIssueCount += diagnostics.length;
-    });
-    updateStatusBar();
-}
-
-function countIssuesBySeverity(severity: vscode.DiagnosticSeverity): number {
-    let count = 0;
-    diagnosticCollection.forEach((uri, diagnostics) => {
-        count += diagnostics.filter(d => d.severity === severity).length;
-    });
-    return count;
-}
-
-// Smart analyzer detection following Go's installation precedence
-async function findAnalyzer(): Promise<string | undefined> {
-    const config = vscode.workspace.getConfiguration('mtlog');
-    const configuredPath = config.get<string>('analyzerPath');
-    
-    // 1. Check explicit setting first
-    if (configuredPath && configuredPath !== 'mtlog-analyzer') {
-        if (fs.existsSync(configuredPath)) {
-            VSCodeMtlogLogger.debug(`Analyzer found at configured path: ${configuredPath}`);
-            return configuredPath;
-        }
-        VSCodeMtlogLogger.warn(`Configured analyzer path not found: ${configuredPath}`, false);
-    }
-    
-    // 2. Try to find in PATH
-    try {
-        const pathResult = execSync('which mtlog-analyzer', { encoding: 'utf8' }).trim();
-        if (pathResult) {
-            VSCodeMtlogLogger.debug(`Analyzer found in PATH: ${pathResult}`);
-            return pathResult;
-        }
-    } catch (e) {
-        // which command failed, continue searching
-    }
-    
-    // For Windows, try where command
-    if (os.platform() === 'win32') {
-        try {
-            const whereResult = execSync('where mtlog-analyzer', { encoding: 'utf8' }).trim().split('\n')[0];
-            if (whereResult) {
-                VSCodeMtlogLogger.debug(`Analyzer found via where: ${whereResult}`);
-                return whereResult;
-            }
-        } catch (e) {
-            // where command failed, continue searching
-        }
-    }
-    
-    // 3. Check Go installation locations following Go's precedence
-    const goLocations: string[] = [];
-    const binaryName = os.platform() === 'win32' ? 'mtlog-analyzer.exe' : 'mtlog-analyzer';
-    
-    // Check GOBIN first
-    const gobin = process.env.GOBIN;
-    if (gobin) {
-        goLocations.push(path.join(gobin, binaryName));
-    }
-    
-    // Check GOPATH/bin
-    const gopath = process.env.GOPATH || path.join(os.homedir(), 'go');
-    goLocations.push(path.join(gopath, 'bin', binaryName));
-    
-    // Check default $HOME/go/bin if different from GOPATH
-    const defaultGoPath = path.join(os.homedir(), 'go', 'bin', binaryName);
-    if (!goLocations.includes(defaultGoPath)) {
-        goLocations.push(defaultGoPath);
-    }
-    
-    // Platform-specific locations
-    if (os.platform() === 'darwin') {
-        // Homebrew locations
-        goLocations.push('/usr/local/bin/mtlog-analyzer');
-        goLocations.push('/opt/homebrew/bin/mtlog-analyzer');
-    }
-    
-    // Check each location
-    for (const location of goLocations) {
-        if (fs.existsSync(location)) {
-            VSCodeMtlogLogger.debug(`Analyzer found at: ${location}`);
-            return location;
-        }
-    }
-    
-    // Log all checked locations for debugging
-    VSCodeMtlogLogger.debug(`Analyzer not found. Checked locations:\n${goLocations.join('\n')}`);
-    return undefined;
-}
-
-async function checkAnalyzerAvailable() {
-    const analyzerPath = await findAnalyzer();
-    
-    if (analyzerPath) {
-        // Update the configuration with the found path for performance
-        const config = vscode.workspace.getConfiguration('mtlog');
-        const currentPath = config.get<string>('analyzerPath');
-        if (currentPath === 'mtlog-analyzer') {
-            // Only update if using default value
-            await config.update('analyzerPath', analyzerPath, vscode.ConfigurationTarget.Global);
-            VSCodeMtlogLogger.info(`Updated analyzer path to: ${analyzerPath}`);
-        }
-        return;
-    }
-    
-    // Show detailed error message with specific paths
-    const gopath = process.env.GOPATH || path.join(os.homedir(), 'go');
-    const expectedPath = path.join(gopath, 'bin', os.platform() === 'win32' ? 'mtlog-analyzer.exe' : 'mtlog-analyzer');
-    
-    const message = `mtlog-analyzer not found in PATH or standard Go locations.\n\nExpected location: ${expectedPath}`;
-    
-    const selection = await vscode.window.showErrorMessage(
-        message,
-        'Install Now',
-        'Configure Path',
-        'Dismiss'
-    );
-    
-    if (selection === 'Install Now') {
-        await installAnalyzer();
-    } else if (selection === 'Configure Path') {
-        await vscode.commands.executeCommand('workbench.action.openSettings', 'mtlog.analyzerPath');
-    }
-}
-
-async function installAnalyzer() {
-    // Check if Go is installed first
-    let goInstalled = false;
-    try {
-        execSync('go version', { encoding: 'utf8' });
-        goInstalled = true;
-    } catch (e) {
-        // Go not found
-    }
-    
-    const options = goInstalled 
-        ? ['Install with Go (recommended)', 'Configure Path Manually', 'Cancel']
-        : ['Configure Path Manually', 'Install Go First', 'Cancel'];
-    
-    const installMethod = await vscode.window.showQuickPick(
-        options,
-        { 
-            placeHolder: goInstalled 
-                ? 'How would you like to install mtlog-analyzer?' 
-                : 'Go is not installed. Please install Go first or configure the path manually.'
-        }
-    );
-    
-    if (!installMethod || installMethod === 'Cancel') {
-        return;
-    }
-    
-    if (installMethod === 'Install Go First') {
-        vscode.env.openExternal(vscode.Uri.parse('https://go.dev/doc/install'));
-        return;
-    }
-    
-    if (installMethod === 'Configure Path Manually') {
-        await vscode.commands.executeCommand('workbench.action.openSettings', 'mtlog.analyzerPath');
-        vscode.window.showInformationMessage(
-            'Please set the full path to mtlog-analyzer in the settings.',
-            'Show Instructions'
-        ).then(selection => {
-            if (selection === 'Show Instructions') {
-                const gopath = process.env.GOPATH || path.join(os.homedir(), 'go');
-                const expectedPath = path.join(gopath, 'bin', os.platform() === 'win32' ? 'mtlog-analyzer.exe' : 'mtlog-analyzer');
-                vscode.window.showInformationMessage(
-                    `After installing with 'go install github.com/willibrandon/mtlog/cmd/mtlog-analyzer@latest', the analyzer will be at:\n\n${expectedPath}`
-                );
-            }
-        });
-        return;
-    }
-    
-    if (installMethod === 'Install with Go (recommended)') {
-        // Show progress notification
-        vscode.window.withProgress({
-            location: vscode.ProgressLocation.Notification,
-            title: "Installing mtlog-analyzer",
-            cancellable: false
-        }, async (progress) => {
-            progress.report({ increment: 0, message: "Running go install..." });
-            
-            try {
-                // Run go install
-                execSync('go install github.com/willibrandon/mtlog/cmd/mtlog-analyzer@latest', {
-                    encoding: 'utf8',
-                    env: { ...process.env }
-                });
-                
-                progress.report({ increment: 100, message: "Installation complete!" });
-                
-                // Try to find the installed analyzer
-                const installedPath = await findAnalyzer();
-                if (installedPath) {
-                    const config = vscode.workspace.getConfiguration('mtlog');
-                    await config.update('analyzerPath', installedPath, vscode.ConfigurationTarget.Global);
-                    
-                    vscode.window.showInformationMessage(
-                        `mtlog-analyzer installed successfully at: ${installedPath}`,
-                        'Reload Window'
-                    ).then(selection => {
-                        if (selection === 'Reload Window') {
-                            vscode.commands.executeCommand('workbench.action.reloadWindow');
-                        }
-                    });
-                } else {
-                    vscode.window.showWarningMessage(
-                        'Installation completed but analyzer not found in expected locations. You may need to reload VS Code.',
-                        'Reload Window'
-                    ).then(selection => {
-                        if (selection === 'Reload Window') {
-                            vscode.commands.executeCommand('workbench.action.reloadWindow');
-                        }
-                    });
-                }
-            } catch (error) {
-                vscode.window.showErrorMessage(
-                    `Failed to install mtlog-analyzer: ${error}. Please check the output for details.`
-                );
-                VSCodeMtlogLogger.error('Installation failed', error as Error);
-            }
-        });
-    }
-}
-
-function showDiagnosticsSummary() {
-    outputChannel.clear();
-    outputChannel.appendLine('=== mtlog-analyzer summary ===');
-    outputChannel.appendLine('');
-    
-    const errorCount = countIssuesBySeverity(vscode.DiagnosticSeverity.Error);
-    const warningCount = countIssuesBySeverity(vscode.DiagnosticSeverity.Warning);
-    const infoCount = countIssuesBySeverity(vscode.DiagnosticSeverity.Information);
-    
-    if (totalIssueCount === 0) {
-        outputChannel.appendLine('✓ No issues found');
-    } else {
-        outputChannel.appendLine(`Found ${totalIssueCount} issue${totalIssueCount !== 1 ? 's' : ''}:`);
-        if (errorCount > 0) outputChannel.appendLine(`  • ${errorCount} error${errorCount !== 1 ? 's' : ''}`);
-        if (warningCount > 0) outputChannel.appendLine(`  • ${warningCount} warning${warningCount !== 1 ? 's' : ''}`);
-        if (infoCount > 0) outputChannel.appendLine(`  • ${infoCount} suggestion${infoCount !== 1 ? 's' : ''}`);
-        
-        outputChannel.appendLine('');
-        outputChannel.appendLine('Issues by file:');
-        
-        const fileIssues: Map<string, vscode.Diagnostic[]> = new Map();
-        diagnosticCollection.forEach((uri, diagnostics) => {
-            if (diagnostics.length > 0) {
-                fileIssues.set(uri.fsPath, [...diagnostics]);
-            }
-        });
-        
-        Array.from(fileIssues.entries())
-            .sort(([a], [b]) => a.localeCompare(b))
-            .forEach(([file, diagnostics]) => {
-                const relPath = vscode.workspace.asRelativePath(file);
-                outputChannel.appendLine(`  ${relPath}: ${diagnostics.length} issue${diagnostics.length !== 1 ? 's' : ''}`);
-            });
-    }
-    
-    outputChannel.appendLine('');
-    outputChannel.appendLine('View all issues in the Problems panel (Ctrl+Shift+M)');
-}
-
-// Code action provider for quick fixes
-class MtlogCodeActionProvider implements vscode.CodeActionProvider {
-    public static readonly providedCodeActionKinds = [
-        vscode.CodeActionKind.QuickFix
-    ];
-
-    provideCodeActions(
-        document: vscode.TextDocument,
-        range: vscode.Range | vscode.Selection,
-        context: vscode.CodeActionContext,
-        token: vscode.CancellationToken
-    ): vscode.CodeAction[] {
-        const actions: vscode.CodeAction[] = [];
-        
-        outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Code actions requested for ${context.diagnostics.length} diagnostics`);
-        
-        // Collect unique diagnostic IDs to avoid duplicates
-        const uniqueDiagnosticIds = new Set<string>();
-        const mtlogDiagnostics: MtlogDiagnostic[] = [];
-        
-        // Process each diagnostic in the current range
-        for (const diagnostic of context.diagnostics) {
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Diagnostic: source=${diagnostic.source}, code=${(diagnostic as any).code}, message=${diagnostic.message}`);
-            
-            if (diagnostic.source !== 'mtlog') continue;
-            
-            const mtlogDiag = diagnostic as MtlogDiagnostic;
-            mtlogDiagnostics.push(mtlogDiag);
-            
-            // Get diagnostic ID from code property (set by the analyzer)
-            const diagnosticId = (diagnostic as any).code;
-            if (diagnosticId) {
-                uniqueDiagnosticIds.add(diagnosticId);
-            }
-        }
-        
-        // Add suppress actions for unique diagnostic IDs
-        for (const diagnosticId of uniqueDiagnosticIds) {
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Adding suppress action for ${diagnosticId}`);
-            const suppressAction = new vscode.CodeAction(
-                `Suppress ${diagnosticId} diagnostic`,
-                vscode.CodeActionKind.QuickFix
-            );
-            suppressAction.command = {
-                command: 'mtlog.suppressDiagnostic',
-                title: 'Suppress diagnostic',
-                arguments: [diagnosticId]
-            };
-            suppressAction.isPreferred = true;
-            actions.push(suppressAction);
-        }
-        
-        // Add analyzer-provided quick fixes for each diagnostic
-        for (const mtlogDiag of mtlogDiagnostics) {
-            // Only use analyzer provided suggested fixes
-            if (mtlogDiag.suggestedFixes && mtlogDiag.suggestedFixes.length > 0) {
-                outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Using ${mtlogDiag.suggestedFixes.length} suggested fixes from analyzer`);
-                for (const suggestedFix of mtlogDiag.suggestedFixes) {
-                    const action = this.createActionFromAnalyzerFix(document, mtlogDiag, suggestedFix);
-                    if (action) {
-                        actions.push(action);
-                    }
-                }
-            }
-        }
-        
-        return actions;
-    }
-    
-    
-    private createActionFromAnalyzerFix(
-        document: vscode.TextDocument,
-        diagnostic: MtlogDiagnostic,
-        suggestedFix: any
-    ): vscode.CodeAction | null {
-        if (!suggestedFix.textEdits || suggestedFix.textEdits.length === 0) {
-            return null;
-        }
-        
-        const action = new vscode.CodeAction(
-            suggestedFix.message || 'Apply fix',
-            vscode.CodeActionKind.QuickFix
-        );
-        
-        action.edit = new vscode.WorkspaceEdit();
-        action.diagnostics = [diagnostic];
-        
-        try {
-            // Apply each text edit from the analyzer
-            for (const edit of suggestedFix.textEdits) {
-                // Parse position from analyzer format (file:line:col)
-                const startParts = edit.pos.split(':');
-                const endParts = edit.end.split(':');
-                
-                if (startParts.length < 3 || endParts.length < 3) {
-                    outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Invalid position format in text edit: ${edit.pos} -> ${edit.end}`);
-                    continue;
-                }
-                
-                const startLine = parseInt(startParts[startParts.length - 2]) - 1; // Convert to 0-based
-                const startCol = parseInt(startParts[startParts.length - 1]) - 1;
-                const endLine = parseInt(endParts[endParts.length - 2]) - 1;
-                const endCol = parseInt(endParts[endParts.length - 1]) - 1;
-                
-                const range = new vscode.Range(
-                    new vscode.Position(startLine, startCol),
-                    new vscode.Position(endLine, endCol)
-                );
-                
-                action.edit.replace(document.uri, range, edit.newText);
-                
-                outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Text edit: ${startLine}:${startCol} -> ${endLine}:${endCol} = "${edit.newText}"`);
-            }
-            
-            // Add command to save and reanalyze after applying the fix
-            action.command = {
-                command: 'mtlog.saveAndAnalyze',
-                title: 'Save and reanalyze'
-            };
-            
-            return action;
-        } catch (e) {
-            outputChannel.appendLine(`[${new Date().toLocaleTimeString()}] Error creating action from analyzer fix: ${e}`);
-            return null;
-        }
-    }
-    
-    private findClosingParen(text: string, startPos: number): number {
-        let parenCount = 0;
-        let inString = false;
-        let stringChar = '';
-        
-        for (let i = startPos; i < text.length; i++) {
-            const char = text[i];
-            
-            // Handle string literals
-            if ((char === '"' || char === '\'' || char === '`') && (i === 0 || text[i-1] !== '\\')) {
-                if (!inString) {
-                    inString = true;
-                    stringChar = char;
-                } else if (char === stringChar) {
-                    inString = false;
-                }
-                continue;
-            }
-            
-            if (!inString) {
-                if (char === '(') parenCount++;
-                else if (char === ')') {
-                    parenCount--;
-                    if (parenCount === 0) return i;
-                }
-            }
-        }
-        
-        return -1;
-    }
-}
-
-
-function getDiagnosticDescription(id: string): string {
-    const descriptions: Record<string, string> = {
-        'MTLOG001': 'Template/argument count mismatch',
-        'MTLOG002': 'Invalid format specifier',
-        'MTLOG003': 'Duplicate property names',
-        'MTLOG004': 'Property naming (PascalCase)',
-        'MTLOG005': 'Missing capturing hints',
-        'MTLOG006': 'Error logging without error value',
-        'MTLOG007': 'Context key constant suggestion',
-        'MTLOG008': 'Dynamic template warning'
-    };
-    return descriptions[id] || 'Unknown diagnostic';
+    cleanupProcesses();
 }
